Rename cache_unit to file_statistics_cache
