+++ datafusion/core/tests/sql/runtime_config.rs
@@ -25,7 +25,7 @@ use datafusion::execution::context::TaskContext;
 use datafusion::prelude::SessionConfig;
 use datafusion_execution::cache::DefaultListFilesCache;
 use datafusion_execution::cache::cache_manager::CacheManagerConfig;
-use datafusion_execution::cache::cache_unit::DefaultFileStatisticsCache;
+use datafusion_execution::cache::file_statistics_cache::DefaultFileStatisticsCache;
 use datafusion_execution::runtime_env::RuntimeEnvBuilder;
 use datafusion_physical_plan::common::collect;
